New Doc2Vec model generator and test suite
This works as very rough classification of topics but suffers from several limitations. The first is the large number of topics (subreddits) makes exact matches almost impossible. Instead, ideally the subreddits listed will be on the same topic as the label (ClimateChange etc.). The steps from here are gathering more specialized data for our purposes (see Issues), and then using a classifier to maximize accuracy.