New Core File
This iteration brings a number of improvements. First, Everything now exists in a function, making it easier to manage. Second, everything is now back in one file as a result. Third, there are now clustering functions set up, as well as plotting functions. Lastly, this iteration is perfectly compatible with the recently uploaded data and model files, allowing contributors to follow along, or do their own analysis with the code.